# Fixes for GCC Renamed 'not' to 'negated', since 'not' is special in C++. Added some includes, etc.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@4665 3db68df2-c116-0410-a063-a993310a9797

+++ source/tools/atlas/AtlasUI/ScenarioEditor/Sections/Environment/LightControl.cpp
@@ -55,7 +55,7 @@ public:
 					float rdotl = rx*lx + ry*ly + rz*lz;
 
 					int diffuse = (int)std::max(0.f, ndotl*128.f);
-					int specular = (int)std::min(255.f, 64.f*pow(std::max(0.f, rdotl), 16.f));
+					int specular = (int)std::min(255.f, 64.f*powf(std::max(0.f, rdotl), 16.f));
 
 					imgData[0] = std::min(64+diffuse+specular, 255);
 					imgData[1] = std::min(48+diffuse+specular, 255);
